@@ -11,11 +11,11 @@
 Gem::Specification.new do |s|
   s.platform    = Gem::Platform::RUBY
   s.name        = 'spree_amazon_payments'
-  s.version     = '3.0.0'
+  s.version     = '3.4.4'
   s.summary     = 'Spree Amazon Payments'
   s.description = ''
 
-  s.required_ruby_version     = '>= 2.1.0'
+  s.required_ruby_version = ['>= 2.3.0', '< 2.5.0']
   s.required_rubygems_version = '>= 1.8.23'
 
   s.author    = 'Amazon Payments'
@@ -23,15 +23,30 @@ Gem::Specification.new do |s|
   s.require_path = 'lib'
   s.requirements << 'none'
 
-  s.add_dependency 'spree_core', '~> 3.0.0'
+  s.add_dependency 'httparty'
+  s.add_dependency 'spree_core', '>= 3.1.0', '< 4.0'
 
-  s.add_development_dependency 'capybara', '~> 2.4'
+  s.add_development_dependency 'appraisal'
+  s.add_development_dependency 'capybara'
+  s.add_development_dependency 'capybara-screenshot'
+  s.add_development_dependency 'codeclimate-test-reporter'
   s.add_development_dependency 'coffee-rails'
   s.add_development_dependency 'database_cleaner'
-  s.add_development_dependency 'factory_girl', '~> 4.4'
+  s.add_development_dependency 'factory_bot'
+  s.add_development_dependency 'factory_bot_rails'
   s.add_development_dependency 'ffaker'
-  s.add_development_dependency 'rspec-rails',  '~> 3.1'
-  s.add_development_dependency 'sass-rails', '~> 4.0.2'
+  s.add_development_dependency 'guard'
+  s.add_development_dependency 'guard-bundler'
+  s.add_development_dependency 'guard-rspec'
+  s.add_development_dependency 'guard-rubocop'
+  s.add_development_dependency 'guard-zeus'
+  s.add_development_dependency 'mysql2'
+  s.add_development_dependency 'pg'
+  s.add_development_dependency 'poltergeist'
+  s.add_development_dependency 'pry-rails'
+  s.add_development_dependency 'rspec-rails'
+  s.add_development_dependency 'rubocop-rspec'
+  s.add_development_dependency 'sass-rails'
   s.add_development_dependency 'selenium-webdriver'
   s.add_development_dependency 'simplecov'
   s.add_development_dependency 'sqlite3'
