@@ -8,7 +8,7 @@
 #
 ##
 class Spree::AmazonCallbackController < ApplicationController
-  skip_before_filter :verify_authenticity_token
+  skip_before_action :verify_authenticity_token
 
   def new
     response = JSON.parse(request.body.read)
